+++ src/Constants/ReducerConstants.ts
@@ -1,3 +1,0 @@
-export const fileNameExtension = 'Reducer'
-
-export const rootFileName = 'Root'