+++ src/Constants/SagasConstants.ts
@@ -1,3 +1,0 @@
-export const fileNameExtension = 'Sagas'
-
-export const rootFileName = 'Root'