+++ src/Controllers/SagasController.ts
@@ -1,6 +1,5 @@
 import { writeFileSync } from 'fs'
 import { SagasFolder } from '../Constants/FolderConstants'
-import { fileNameExtension } from '../Constants/SagasConstants'
 import ParsedProperty from '../Models/ParsedProperty'
 import {
     capitalize,
@@ -11,6 +10,9 @@ import AbstractFluxController from './AbstractFluxController'
 import ActionCreatorController from './ActionCreatorController'
 import ActionTypeController from './ActionTypeController'
 import ServiceController from './ServiceController'
+
+export const fileNameExtension = 'Sagas'
+export const rootFileName = 'Root'
 
 export default class SagasController extends AbstractFluxController {
     private sagaMainFunction: string
@@ -28,6 +30,10 @@ export default class SagasController extends AbstractFluxController {
         return super.createFile(modelName, fluxExtension)
     }
 
+    public serviceImports(serviceName: string): string {
+        return `    ${serviceName},`
+    }
+
     public generateSagaImports(
         actionTypeController: ActionTypeController,
         actionCreatorController: ActionCreatorController,
@@ -37,13 +43,15 @@ export default class SagasController extends AbstractFluxController {
             'import { put, call, all, fork, takeLatest } from \'redux-saga/effects\''
         const actionCreatorImport = `import * as actionCreators from '@/${actionCreatorController.getFolderName()}/${actionCreatorController
             .getFileName()
-            .replace('.ts', '')}'`
+            .replace(/\.ts$/, '')}'`
         const actionTypeImport = `import * as actionTypes from '@/${actionTypeController.getFolderName()}/${actionTypeController
             .getFileName()
-            .replace('.ts', '')}'`
-        const serviceImport = `import {\n${serviceController.getServiceName()},\n} from '@/${serviceController.getFolderName()}/${serviceController
+            .replace(/\.ts$/, '')}'`
+        const serviceImport = `import {\n${this.serviceImports(
+            serviceController.getServiceName()
+        )}\n} from '@/${serviceController.getFolderName()}/${serviceController
             .getFileName()
-            .replace('.ts', '')}'`
+            .replace(/\.ts$/, '')}'`
 
         // TODO Set this variable to be configurable by the user
         const errorHandlerImport =
@@ -89,10 +97,9 @@ export default class SagasController extends AbstractFluxController {
         const sagaFunction = `function* ${sagaFunctionName}({\n${params}\n\
 }: actionTypes.${actionTypeController.getActionTypeInterfacesNames()[0]}) {\n\
     const params = {\n${propertiesToReturnAction(propertiesSend, '        ')}\n\
-    }\n\
+    }\n\n\
     try {\n\
-        const { data } = yield call(${serviceController.getServiceName()}, params)\n\
-        \n\
+        const { data } = yield call(${serviceController.getServiceName()}, params)\n\\n\
         yield put(actionCreators.${
     actionTypeController.getActionTypeIdentifiers()[0]
 }Success(${propertiesSuccess