+++ src/fluxCreator/createActionCreator.ts
@@ -25,7 +25,24 @@ export default function createActionCreator(
     ]
 
     if (data.length) {
-        // TODO Append to file and check if it doesn't already exists
+        actionCreatorController.addLine('\n')
+
+        actionTypeController.getActionTypeNames().forEach((el, i) => {
+            actionCreatorController.addActionCreator(
+                actionTypeController.getActionTypeInterfacesNames()[i],
+                el,
+                actionTypeController.getActionTypeIdentifiers()[i],
+                properties[i].sort((a, b) =>
+                    a.optional && !b.optional
+                        ? -1
+                        : a.optional && b.optional
+                            ? 0
+                            : 1
+                )
+            )
+        })
+
+        actionCreatorController.appendFile(fd)
     } else {
         actionCreatorController.generateImports(
             actionTypeController,
@@ -37,8 +54,12 @@ export default function createActionCreator(
                 actionTypeController.getActionTypeInterfacesNames()[i],
                 el,
                 actionTypeController.getActionTypeIdentifiers()[i],
-                properties[i].sort(
-                    (a, b) => (a.optional && !b.optional && -1) || 1
+                properties[i].sort((a, b) =>
+                    a.optional && !b.optional
+                        ? -1
+                        : a.optional && b.optional
+                            ? 0
+                            : 1
                 )
             )
         })