+++ src/fluxCreator/createReducer.ts
@@ -1,7 +1,8 @@
 import { closeSync, ftruncateSync, readFileSync } from 'fs'
-import { rootFileName } from '../Constants/ReducerConstants'
 import ActionTypeController from '../Controllers/ActionTypeController'
-import ReducerController from '../Controllers/ReducerController'
+import ReducerController, {
+    rootFileName,
+} from '../Controllers/ReducerController'
 import ParsedModel from '../Models/ParsedModel'
 import ParsedProperty from '../Models/ParsedProperty'
 
@@ -61,7 +62,15 @@ export default function createReducer(
     ]
 
     if (data.length) {
-        // TODO Append to file
+        const reducers: string[] = []
+
+        actionTypeController.getActionTypeNames().forEach((el, i) => {
+            reducers.push(
+                reducerController.generateReducer(el, allProperties[i])
+            )
+        })
+
+        reducerController.appendReducers(fd, reducers)
     } else {
         reducerController.generateImports(
             actionTypeController,